refactor(renderer): split state.rs by audience
state.rs mixed three audiences — the public TocEntry output type,
walker-private scratch (HeadingAccumulator, TableState), and pure
utilities (escape_html, slugify). Untangled into focused modules:

  - util.rs gains escape_html (pub) and slugify (pub(crate))
  - toc.rs holds TocEntry, CompletedHeading, HeadingAccumulator
  - table.rs holds TableState
  - state.rs deleted

Tests move with their items. No semantic change: rw_renderer::TocEntry
and rw_renderer::escape_html keep their paths, so rw-kroki (the only
external consumer of escape_html) compiles untouched.

Also fixes a stale intra-doc link carried forward from state.rs:
`[`Scope::Heading`](crate::renderer::Scope)` → `crate::scope::Scope`.
Scope moved to its own module in an earlier refactor; the link was a
hard error under `-D rustdoc::broken_intra_doc_links`.
