+++ crates/rw-renderer/src/html.rs
@@ -10,7 +10,7 @@ use std::borrow::Cow;
 use std::fmt::Write;
 
 use crate::backend::{AlertKind, RenderBackend};
-use crate::state::escape_html;
+use crate::util::escape_html;
 
 // SVG icons for alerts (GitHub Octicons-style, 16x16)
 const SVG_INFO: &str = r#"<svg class="alert-icon" viewBox="0 0 16 16" width="16" height="16" aria-hidden="true"><path d="M0 8a8 8 0 1 1 16 0A8 8 0 0 1 0 8Zm8-6.5a6.5 6.5 0 1 0 0 13 6.5 6.5 0 0 0 0-13ZM6.5 7.75A.75.75 0 0 1 7.25 7h1a.75.75 0 0 1 .75.75v2.75h.25a.75.75 0 0 1 0 1.5h-2a.75.75 0 0 1 0-1.5h.25v-2h-.25a.75.75 0 0 1-.75-.75ZM8 6a1 1 0 1 1 0-2 1 1 0 0 1 0 2Z"></path></svg>"#;
