+++ crates/rw-renderer/src/toc.rs
@@ -1,62 +1,14 @@
-//! Shared state structs for markdown rendering.
+//! Table of contents output type and heading accumulator.
 //!
-//! These structs track context during event processing and are shared
-//! between HTML and Confluence backends.
+//! [`TocEntry`] is the public output type collected in
+//! [`RenderResult::toc`](crate::RenderResult::toc).
+//! [`HeadingAccumulator`] is walker-private scratch that tracks
+//! cross-heading state (title, TOC entries, id-counts, and the
+//! "have we seen the first H1?" flag) across an entire document render.
 
 use std::collections::HashMap;
 
-use pulldown_cmark::Alignment;
-
-/// State for tracking table rendering.
-#[derive(Default)]
-pub(crate) struct TableState {
-    /// Whether we're inside the table header row.
-    in_head: bool,
-    /// Column alignments for current table.
-    alignments: Vec<Alignment>,
-    /// Current column index in table row.
-    cell_index: usize,
-}
-
-impl TableState {
-    /// Start a new table with column alignments.
-    pub fn start(&mut self, alignments: Vec<Alignment>) {
-        self.alignments = alignments;
-        self.in_head = false;
-        self.cell_index = 0;
-    }
-
-    /// Start the table header row.
-    pub fn start_head(&mut self) {
-        self.in_head = true;
-        self.cell_index = 0;
-    }
-
-    /// End the table header row.
-    pub fn end_head(&mut self) {
-        self.in_head = false;
-    }
-
-    /// Start a new table row.
-    pub fn start_row(&mut self) {
-        self.cell_index = 0;
-    }
-
-    /// Move to the next cell.
-    pub fn next_cell(&mut self) {
-        self.cell_index += 1;
-    }
-
-    /// Check if we're in the table header.
-    pub fn is_in_head(&self) -> bool {
-        self.in_head
-    }
-
-    /// Get the alignment for the current cell.
-    pub fn current_alignment(&self) -> Option<Alignment> {
-        self.alignments.get(self.cell_index).copied()
-    }
-}
+use crate::util::slugify;
 
 /// A single heading in the table of contents.
 ///
@@ -118,7 +70,7 @@ pub(crate) struct CompletedHeading {
 ///
 /// Per-heading state (the heading's level, plain-text shadow, formatted
 /// HTML body, and the `in_first_h1` flag) lives in
-/// [`Scope::Heading`](crate::renderer::Scope) — *not* here. The
+/// [`Scope::Heading`](crate::scope::Scope) — *not* here. The
 /// accumulator is consulted at `Tag::Heading` start (to decide whether
 /// the heading is the skipped Confluence first H1) and at
 /// `TagEnd::Heading` (to capture the title and/or emit the heading).
@@ -235,118 +187,9 @@ impl HeadingAccumulator {
     }
 }
 
-/// Convert text to URL-safe slug.
-///
-/// Converts to lowercase, replaces whitespace/dashes/underscores with single dashes,
-/// and removes other non-alphanumeric characters. Preserves non-Latin Unicode characters
-/// (Cyrillic, CJK, etc.) following GitHub-style heading ID generation.
-#[must_use]
-fn slugify(text: &str) -> String {
-    let mut result = String::new();
-    let mut last_was_dash = true; // Prevents leading dash
-
-    for c in text.trim().chars() {
-        if c.is_alphanumeric() {
-            for lc in c.to_lowercase() {
-                result.push(lc);
-            }
-            last_was_dash = false;
-        } else if !last_was_dash && (c.is_whitespace() || c == '-' || c == '_') {
-            result.push('-');
-            last_was_dash = true;
-        }
-    }
-
-    // Remove trailing dash if present
-    if result.ends_with('-') {
-        result.pop();
-    }
-
-    result
-}
-
-/// Escapes the five HTML special characters (`&`, `<`, `>`, `"`, `'`).
-///
-/// # Examples
-///
-/// ```
-/// use rw_renderer::escape_html;
-///
-/// assert_eq!(escape_html("<script>"), "&lt;script&gt;");
-/// assert_eq!(escape_html(r#"a "b" & 'c'"#), "a &quot;b&quot; &amp; &#x27;c&#x27;");
-/// ```
-#[must_use]
-pub fn escape_html(s: &str) -> String {
-    let mut result = String::with_capacity(s.len());
-    for c in s.chars() {
-        match c {
-            '&' => result.push_str("&amp;"),
-            '<' => result.push_str("&lt;"),
-            '>' => result.push_str("&gt;"),
-            '"' => result.push_str("&quot;"),
-            '\'' => result.push_str("&#x27;"),
-            _ => result.push(c),
-        }
-    }
-    result
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_slugify() {
-        assert_eq!(slugify("Hello World"), "hello-world");
-        assert_eq!(slugify("What's New?"), "whats-new");
-        assert_eq!(slugify("  Spaces  "), "spaces");
-        assert_eq!(slugify("Multiple   Spaces"), "multiple-spaces");
-        assert_eq!(slugify("kebab-case"), "kebab-case");
-        assert_eq!(slugify("snake_case"), "snake-case");
-    }
-
-    #[test]
-    fn test_slugify_non_latin() {
-        // Cyrillic
-        assert_eq!(slugify("Привет мир"), "привет-мир");
-        // Chinese
-        assert_eq!(slugify("你好世界"), "你好世界");
-        // Japanese
-        assert_eq!(slugify("こんにちは世界"), "こんにちは世界");
-        // Mixed Latin and non-Latin
-        assert_eq!(slugify("Hello Привет"), "hello-привет");
-        // Non-Latin with punctuation
-        assert_eq!(slugify("Привет, мир!"), "привет-мир");
-        // Fully non-Latin should NOT produce empty string
-        assert!(!slugify("Заголовок").is_empty());
-    }
-
-    #[test]
-    fn test_escape_html() {
-        assert_eq!(escape_html("<script>"), "&lt;script&gt;");
-        assert_eq!(escape_html("a & b"), "a &amp; b");
-        assert_eq!(escape_html(r#""quoted""#), "&quot;quoted&quot;");
-        assert_eq!(escape_html("it's"), "it&#x27;s");
-    }
-
-    #[test]
-    fn test_table_state() {
-        let mut state = TableState::default();
-        state.start(vec![Alignment::Left, Alignment::Center, Alignment::Right]);
-
-        state.start_head();
-        assert!(state.is_in_head());
-        assert_eq!(state.current_alignment(), Some(Alignment::Left));
-
-        state.next_cell();
-        assert_eq!(state.current_alignment(), Some(Alignment::Center));
-
-        state.next_cell();
-        assert_eq!(state.current_alignment(), Some(Alignment::Right));
-
-        state.end_head();
-        assert!(!state.is_in_head());
-    }
 
     #[test]
     fn test_heading_accumulator_html_mode() {
